Added operator += for all existing iterators
This was required as for MacOS a compile error would be triggered for the sort algorithm if this operator is missing. This happened obviously for the non-const iterators however for consistency reasons the operator has been implemented for the const ones too. It is also planned to implement a -= operator