Updated the conversion function to C-Style array
It was required to simplify the function by passing a contiguous unidimensional array to the user without any reference to extra capacity. The array would contain the exact elements of the matrix in row order. Another reason for this change was a safer usage of the pointers (returned pointer to be used for de-allocating the raw array)