Modified the const-ness of iterator returning functions
It was required that the non-const iterator functions (e.g. the ones returning a ZIterator) don't have the const keyword. This is similar to the std::vector iterator behavior and was required in order to avoid const property violation (e.g. when having a const Matrix& argument a ZIterator would work because it had the const keyword)