Removed the Z/NIterator test for isValidWithMatrix()
This method has been created mainly as a helper function for testing purposes so no dedicated test should be created for checking it. Also there was a risk that in the current test some of the cases might exhibit random behavior (i.e. the pointer of the re-allocated memory for a matrix might contain the same or a different address from the one of the previously allocated memory)