Created empty diagonal iterator for all concerned types
This was required in order to be able to correctly use data-driven testing. On some systems it is required that the object has a default constructor in order to be included in the data driven testing table. To be note is that an empty diagonal iterator is not compatible with any matrix (empty or not)