Implemented distributed capacity for Matrix
It was required to have available capacity for rows and columns on both
sides of the used capacity space, i.e. "top/bottom" available capacity
for rows, respectively "left/right" available capacity for columns. This
was needed in order to optimize the process of inserting rows and/or
columns, for example for being able to prepend them without costly
move/copy operations

Also as part of this commit, some unnecessary methods have been removed:
- isValidWithMatrix() for iterators: no longer applicable due to the
  capacity paradigm change -> the user should ensure instead each iterator
is used in a correct manner and with the right container
- specific Matrix methods: swapRowColumn(), swapItems(), setAllItemsToValue(), copy() -> all these could be implemented with iterators and keeping them would have continued to pollute the library code