Implemented max dimensions in constructors and assignment
It was required to setup a maximum allowed number of rows and columns when initializing a Matrix. Due to the fact that the size_type is uint32_t and the diff_type is int64_t an insurance was required that no overflow occurs in the positive space of the diff_type when setting up the matrix dimensions. In order to achieve this the maximum number of rows and columns has been set to 31 out of the 32 bits of the unsigned size_type integer (uint32_t), respectively 7 out of the 8 bits of the shorter version (uint16_t). The shorter version was required in order to prevent an excessively long test run but the tests have been adjusted to ensure both cases are accomodated.

In addition to limiting the size of the created Matrixes a capped capacity was also needed as it doesn't make sense to allocate a higher capacity than the maximum allowed size.

The restrictions have been extended to the assignment operators as well.

TODO:
- extend restrictions to other operations like resizing, inserting rows and columns and in general wherever applicable
- update README.md with these specifications