Implemented resize/reserve separation
It was required to implement a separation of concerns for the original
resize() methods which were responsible both for modifying size and
changing capacity. This was a violation of the SOLID principles, namely
of the single responsibility concept.

For this reason it was required to
make following change:
- resize() is only responsible for changing the size of the matrix.
  Capacity would only get adjusted to ensure size fitting within
capacity.
- reserve is only responsible for changing capacity. Size always stays
  the same and capacity cannot go below the given size.

Another benefit of this change was the opportunity to use the same name
for both resize() methods which originally had to be altered due to the
default values of the capacity parameters (which used to cause some
compiling errors for integer matrixes)