Updated the transpose() method and its tests
It was required to simplify the transpose() method so it can only be
applied to the current matrix. This also meant simplifying the tests. An
upper bound related to max dimension/capacity also had to be set (similarly to
other methods). Some additional tests were added to exercise this upper bound.